Replaced useLayoutEffect with useEffect
```
Warning: useLayoutEffect does nothing on the server, because its effect cannot be encoded into the server renderer's output format. This will lead to a mismatch between the initial, non-hydrated UI and the intended UI. To avoid this, useLayoutEffect should only be used in components that render exclusively on the client. See https://reactjs.org/link/uselayouteffect-ssr for common fixes.
    at In (C:\web\app\node_modules\tunnel-rat\dist\index.cjs.js:17:7)
    at Suspense
    at div
    at C:\web\app\node_modules\@emotion\react\dist\emotion-element-ae8cc4ba.cjs.dev.js:71:25
```